+++ packages/@vuepress/theme-default/src/node/resolveContainerPluginOptions.ts
@@ -4,7 +4,7 @@ import type { DefaultThemeOptions } from '../types'
 /**
  * Resolve options for @vuepress/plugin-container
  *
- * For custom blocks default title
+ * For custom containers default title
  */
 export const resolveContainerPluginOptions = (
   options: DefaultThemeOptions,
@@ -36,7 +36,7 @@ export const resolveContainerPluginOptions = (
 /**
  * Resolve options for @vuepress/plugin-container
  *
- * For details block
+ * For details container
  */
 export const resolveContainerPluginOptionsForDetails = (
   options: DefaultThemeOptions
@@ -48,9 +48,47 @@ export const resolveContainerPluginOptionsForDetails = (
   return {
     type: 'details',
     before: (info) =>
-      `<details class="custom-block details">${
+      `<details class="custom-container details">${
         info ? `<summary>${info}</summary>` : ''
       }\n`,
     after: () => '</details>\n',
+  }
+}
+
+/**
+ * Resolve options for @vuepress/plugin-container
+ *
+ * For code-group container
+ */
+export const resolveContainerPluginOptionsForCodeGroup = (
+  options: DefaultThemeOptions
+): ContainerPluginOptions | boolean => {
+  if (options.themePlugins?.container?.codeGroup === false) {
+    return false
+  }
+
+  return {
+    type: 'code-group',
+    before: () => `<CodeGroup>\n`,
+    after: () => '</CodeGroup>\n',
+  }
+}
+
+/**
+ * Resolve options for @vuepress/plugin-container
+ *
+ * For code-group-item block
+ */
+export const resolveContainerPluginOptionsForCodeGroupItem = (
+  options: DefaultThemeOptions
+): ContainerPluginOptions | boolean => {
+  if (options.themePlugins?.container?.codeGroupItem === false) {
+    return false
+  }
+
+  return {
+    type: 'code-group-item',
+    before: (info) => `<CodeGroupItem title="${info}">\n`,
+    after: () => '</CodeGroupItem>\n',
   }
 }