+++ packages/@vuepress/plugin-pwa-popup/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-alpha.19](https://github.com/vuepress/vuepress-next/compare/v2.0.0-alpha.18...v2.0.0-alpha.19) (2021-01-24)
+
+**Note:** Version bump only for package @vuepress/plugin-pwa-popup
+
+
+
+
+
+# [2.0.0-alpha.18](https://github.com/vuepress/vuepress-next/compare/v2.0.0-alpha.17...v2.0.0-alpha.18) (2021-01-17)
+
+**Note:** Version bump only for package @vuepress/plugin-pwa-popup
+
+
+
+
+
+# [2.0.0-alpha.17](https://github.com/vuepress/vuepress-next/compare/v2.0.0-alpha.16...v2.0.0-alpha.17) (2021-01-13)
+
+**Note:** Version bump only for package @vuepress/plugin-pwa-popup
+
+
+
+
+
 # [2.0.0-alpha.16](https://github.com/vuepress/vuepress-next/compare/v2.0.0-alpha.15...v2.0.0-alpha.16) (2021-01-11)
 
 **Note:** Version bump only for package @vuepress/plugin-pwa-popup